詳細・一覧・データ完了

+++ db/migrate/20210705035512_devise_create_users.rb
@@ -38,9 +38,9 @@ class DeviseCreateUsers < ActiveRecord::Migration[5.2]
       t.string :email
       t.string :profile_image_id
       t.string :introduction
-      t.integer :height
-      t.integer :weight
-      t.integer :bmi
+      t.float :height
+      t.float :weight
+      t.float :bmi
       t.integer :sex
       t.integer :age
       t.boolean :is_valid, null: false, default: true
